organize child components of Views Sources
